5_Arduino/VolumeCotrol/requirements.txt
comtypes==1.1.14
psutil==5.9.2
pycaw==20220416
pyserial==3.5
